# src/main/java/com/example/demo/domain/model/ConfluencePage.java
package com.example.demo.domain.model;


public class ConfluencePage {
    private String id;
    private String title;
    private String content;
    private String url;

    public ConfluencePage() {

    }

    public String getId() {
        return id;
    }

    public String getTitle() {
        return title;
    }

    public String getContent() {
        return content;
    }

    public String getUrl() {
        return url;
    }

    public void setId(String id) {
        this.id = id;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public void setContent(String content) {
        this.content = content;
    }

    public void setUrl(String url) {
        this.url = url;
    }
}