# src/main/java/com/example/demo/domain/model/CopilotQuery.java
package com.example.demo.domain.model;


public class CopilotQuery {
    private String question;

    public String getQuestion() {
        return question;
    }

    public void setQuestion(String question) {
        this.question = question;
    }
}